Refactor teeth visual, update factories, and rename views
Introduces a TeethVisual component and updates usage throughout the app, including controllers and Blade views. Refactors Entry model to provide a teeth_list accessor. Updates factories for more realistic data and changes relevant database columns to integer types. Renames Blade view files and their references to use kebab-case for consistency.

+++ database/migrations/2024_03_03_163108_entries.php
@@ -14,11 +14,11 @@ return new class extends Migration {
             $table->id();
             $table->date('date');
             $table->string('teeth');
-            $table->decimal('amount');
-            $table->decimal('unit_price');
+            $table->integer('amount');
+            $table->integer('unit_price');
             $table->decimal('discount');
-            $table->decimal('price');
-            $table->decimal('cost');
+            $table->integer('price');
+            $table->integer('cost');
             $table->foreignId('customer_id')->constrained('customers')->onUpdate('cascade')->onDelete('cascade');
             $table->foreignId('item_id')->constrained('items')->onUpdate('cascade')->onDelete('cascade');
             $table->timestamps();
